+++ Tests/PowerSyncTests/Implementation/DatabaseImplementationTests.swift
@@ -23,4 +23,23 @@ struct DatabaseImplementationTests {
         try await db.close(deleteDatabase: true)
         #expect(description.contains("attempt to write a readonly database"))
     }
+    
+    @Test func canUseConnectionInCallback() async throws {
+        let db = PowerSyncDatabase(
+            schema: Schema(),
+            dbFilename: ":memory:",
+            logger: DefaultLogger()
+        )
+        
+        let results = try await db.readTransaction { tx in
+            try tx.getAll(sql: "VALUES (1), (2), (3)", parameters: []) { outerCursor in
+                let outerValue = try outerCursor.getInt64(index: 0)
+                return try tx.get(sql: "SELECT 2 * ?", parameters: [outerValue]) { innerCursor in
+                    try innerCursor.getInt64(index: 0)
+                }
+            }
+        }
+        try #require(results == [2, 4, 6])
+        try await db.close()
+    }
 }